+++ src/App.tsx
@@ -1,7 +1,5 @@
 import Root from "./Root"
 
-import './css/app.css'
-
 function App() {
   return (
     <Root />