+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased (0.8.0)
+- Added support to report on when health API endpoint is not available in navbar and on health dashboard
+- Fixed clickable area of navbar menu items
+- Added support to hard code links for projects/venues
+- Added health check URL to health dashboard [#40](https://github.com/unity-sds/unity-ui/issues/40)
+- Fixed health dashboard column resizer element z-index
+- Added cards to homepage [#38](https://github.com/unity-sds/unity-ui/issues/38)
+
+## [0.7.0] 2024-09-27
+- Updated node version lts/iron
+- Added ability to fetch health information from Health API Endpoint for Health Dashboard
+- Removed cognito authentication
+- Added authorization via httpd proxy
+- Removed unneeded navigation links
+- Removed job monitoring and new job features
+- Added progress bar to be displayed when health information is being fetched
+- Various code cleanup and optimization activities
+
+## [0.6.0] (unreleased)
+- Added Health Dashboard using static JSON file containing example health information. [#29](https://github.com/unity-sds/unity-ui/issues/29)
+- Updated navbar menu so that it is (partially) dynamic. External UI information is loaded via the Health JSON file. [#20](https://github.com/unity-sds/unity-ui/issues/20)
+
 ## [0.5.0] 2024-04-15
 - Updated Navbar CSS styling to match Figma designs [#5](https://github.com/unity-sds/unity-ui/issues/5)
 - Added CI/CD workflow to build application as a docker image. [#21](https://github.com/unity-sds/unity-ui/issues/21)